+++ core/src/main/java/io/cucumber/core/options/CurlOption.java
@@ -1,5 +1,8 @@
 package io.cucumber.core.options;
 
+import java.net.InetSocketAddress;
+import java.net.Proxy;
+import java.net.Proxy.Type;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.AbstractMap.SimpleEntry;
@@ -7,16 +10,19 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map.Entry;
 
+import static java.net.Proxy.NO_PROXY;
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
 
 public final class CurlOption {
 
+    private final Proxy proxy;
     private final URI uri;
     private final HttpMethod method;
     private final List<Entry<String, String>> headers;
 
-    private CurlOption(HttpMethod method, URI uri, List<Entry<String, String>> headers) {
+    private CurlOption(Proxy proxy, HttpMethod method, URI uri, List<Entry<String, String>> headers) {
+        this.proxy = requireNonNull(proxy);
         this.uri = requireNonNull(uri);
         this.method = requireNonNull(method);
         this.headers = requireNonNull(headers);
@@ -24,15 +30,16 @@ public final class CurlOption {
 
     @SafeVarargs
     public static CurlOption create(HttpMethod method, URI uri, Entry<String, String>... headers) {
-        return new CurlOption(method, uri, asList(headers));
+        return new CurlOption(NO_PROXY, method, uri, asList(headers));
     }
 
     public static CurlOption parse(String cmdLine) {
         List<String> args = ShellWords.parse(cmdLine);
 
-        String url = null;
+        URI url = null;
         HttpMethod method = HttpMethod.PUT;
         List<Entry<String, String>> headers = new ArrayList<>();
+        Proxy proxy = NO_PROXY;
 
         while (!args.isEmpty()) {
             String arg = args.remove(0);
@@ -43,22 +50,64 @@ public final class CurlOption {
                 String headerArg = removeArgFor(arg, args);
                 SimpleEntry<String, String> e = parseHeader(headerArg);
                 headers.add(e);
+            } else if (arg.equals("-x")) {
+                String proxyArg = removeArgFor(arg, args);
+                proxy = parseProxy(proxyArg);
             } else {
                 if (url != null) {
                     throw new IllegalArgumentException("'" + cmdLine + "' was not a valid curl command");
                 }
-                url = arg;
+                url = parseUrl(arg);
             }
         }
 
         if (url == null) {
             throw new IllegalArgumentException("'" + cmdLine + "' was not a valid curl command");
         }
+        return new CurlOption(proxy, method, url, headers);
+    }
+
+    private static URI parseUrl(String arg) {
         try {
-            return new CurlOption(method, new URI(url), headers);
+            return new URI(arg);
         } catch (URISyntaxException e) {
-            throw new IllegalArgumentException(e);
+            throw new IllegalArgumentException("'" + arg + "' was not a valid url", e);
         }
+    }
+
+    private static Proxy parseProxy(String arg) {
+        URI url;
+        try {
+            url = new URI(arg);
+        } catch (URISyntaxException e) {
+            throw new IllegalArgumentException("'" + arg + "' was not a valid proxy address", e);
+        }
+
+        String protocol = url.getScheme();
+        if (protocol == null) {
+            throw new IllegalArgumentException("'" + arg + "' did not have a valid proxy protocol");
+        }
+
+        Proxy.Type type;
+        if (protocol.equalsIgnoreCase("http") || protocol.equalsIgnoreCase("https")) {
+            type = Type.HTTP;
+        } else if (protocol.equalsIgnoreCase("socks")) {
+            type = Type.SOCKS;
+        } else {
+            throw new IllegalArgumentException("'" + arg + "' did not have a valid proxy protocol");
+        }
+
+        String host = url.getHost();
+        if (host == null) {
+            throw new IllegalArgumentException("'" + arg + "' did not have a valid proxy host");
+        }
+
+        int port = url.getPort();
+        if (port == -1) {
+            throw new IllegalArgumentException("'" + arg + "' did not have a valid proxy port");
+        }
+
+        return new Proxy(type, new InetSocketAddress(host, port));
     }
 
     private static String removeArgFor(String arg, List<String> args) {
@@ -86,6 +135,10 @@ public final class CurlOption {
 
     public URI getUri() {
         return uri;
+    }
+
+    public Proxy getProxy() {
+        return proxy;
     }
 
     public enum HttpMethod {